Maybe it works

+++ src/main.cpp
@@ -8,15 +8,15 @@ int main()
     std::cin >> n;
     assert (std::cin.good());
 
-    double* data = new double[n * n]{};
+    long double* data = new long double[n * n]{};
     for (size_t i = 0; i < n * n; i++)
     {
         // std::cerr << i << std::endl;
         std::cin >> data[i];
         assert (std::cin.good());
     }
-    matrix::Matrix<double> M {n, data};
+    matrix::Matrix<long double> M {n, data};
     // std::cerr << "Matrix was made\n";
     delete[] data;
-    std::cout << static_cast<int>(M.det());
+    std::cout << static_cast<long long>(M.det());
 }
